refactor: eliminate barrel exports and fix TypeScript issues
This commit completes Priority 0 Item #1 from the React refactor plan:
eliminating all barrel exports (index.ts files) to improve code traversal
and follow React coding standards.

Changes:
- Remove all barrel export files (hooks/index.ts, stores/index.ts, daemon/index.ts)
- Update 15+ files to use direct imports instead of barrel exports
- Fix TypeScript errors related to missing HLD SDK module resolution
- Fix implicit 'any' parameter types in formatToolResult.tsx
- Fix type union issues in daemon http-client.ts
- Add proper type assertions in useSessionFilter.ts

Verification:
✅ Format check passed
✅ Lint check passed
✅ Type checking passed
✅ All barrel export usage eliminated

Impact:
- Improved code traversal depth (React standard principle #1)
- Better tree-shaking and build performance
- More explicit dependencies
- Eliminated potential circular dependency issues
