+++ .simonsays/.gitignore
@@ -1,0 +1,1 @@
+claude_output.jsonl
