+++ .simonsays/ralph.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+while :; do
+  ./.simonsays/sync.sh
+  echo -e "===SLEEP===\n===SLEEP===\n"; echo 'looping';
+  sleep 10;
+done
