+++ humanlayer-wui/src/components/FuzzySearchInput.tsx
@@ -8,7 +8,7 @@ import { Popover, PopoverAnchor, PopoverContent } from './ui/popover'
 import { Command, CommandEmpty, CommandGroup, CommandItem, CommandList } from './ui/command'
 import { ArrowDownUp, FileWarning, Clock } from 'lucide-react'
 import { useHotkeys } from 'react-hotkeys-hook'
-import type { RecentPath } from '@/lib/daemon'
+import type { RecentPath } from '@/lib/daemon/types'
 
 interface SearchInputProps {
   value?: string
