+++ humanlayer-wui/src/components/internal/SessionDetail/SessionDetail.tsx
@@ -4,7 +4,8 @@ import { toast } from 'sonner'
 
 import { ConversationEvent, Session, ApprovalStatus, SessionStatus } from '@/lib/daemon/types'
 import { Card, CardContent } from '@/components/ui/card'
-import { useConversation, useKeyboardNavigationProtection } from '@/hooks'
+import { useConversation } from '@/hooks/useConversation'
+import { useKeyboardNavigationProtection } from '@/hooks/useKeyboardNavigationProtection'
 import { ChevronDown, Archive, Pencil } from 'lucide-react'
 import { getStatusTextClass } from '@/utils/component-utils'
 import { renderSessionStatus } from '@/utils/sessionStatus'
