--- humanlayer-wui/src/stores/index.ts
@@ -1,5 +1,0 @@
-export { createRealAppStore, createDemoAppStore } from './appStore'
-export type { AppState } from './appStore'
-export { DemoStoreProvider } from './DemoStoreProvider'
-export { DemoAnimator } from './DemoAnimator'
-export type { AnimationStep } from './DemoAnimator'
